requirements-termux.txt
# FastAPI e Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic>=2.10.6
pydantic-settings>=2.7.1

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# Autenticação e Segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Rate Limiting
slowapi==0.1.9

# Cache - DESATIVADO NO TERMUX
# redis==5.0.1
# hiredis==3.3.0

# Logging
loguru==0.7.2

# WebSocket e Conexões
websockets==12.0
aiofiles==23.2.1

# HTTP Requests
requests==2.31.0
httpx~=0.25.2

# Utilitários
python-dotenv==1.0.0
pytz==2023.3

# Telegram (Notificações)
python-telegram-bot==20.7

# Google Sheets (desativado - ngrok removido)
#gspread==5.12.0
#oauth2client==4.1.3

# Testing (opcional)
pytest==7.4.4
pytest-asyncio==0.23.3

# Outros
orjson==3.11.7

# REMOVIDOS PARA TERMUX (requerem compilação pesada):
# numpy==2.3.3
# pandas==2.3.3
# bcrypt==4.1.2
